% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\docType{methods}
\name{ex}
\alias{ex}
\title{Examine the first elements of an object}
\usage{
ex(x, ...)
}
\arguments{
\item{x}{An object to examine}

\item{...}{Arguments to \code{head()} or \code{ex.data.frame()} (See details)}
}
\value{
A data.frame
}
\description{
Examine the first elements of an object
}
\details{
This function is intended to provide a look at an R object. For a data.frame,
it will return by default the first 6 rows and columns. For other non-list
objects, the function defaults to whatever \code{head()} returns.
For list objects (that aren't data.frames), the function will recursively
apply itself to all objects in the list.
}
\examples{
test <- list(iris, letters, list(mtcars, LETTERS))
ex(test)
}
