% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplify.R
\name{examplify}
\alias{examplify}
\title{See an example of a data.frame}
\usage{
examplify(x, rownum = 6, colnum = 6)
}
\arguments{
\item{x}{A data.frame to subset}

\item{rownum}{The number of rows to select}

\item{colnum}{The number of columns to select}
}
\value{
A data.frame
}
\description{
See an example of a data.frame
}
\details{
This function is intended just to provide a glimpse into a data.frame, 
returning the first \code{rownum} rows and \code{colnum} columns
}
\examples{
examplify(mtcars)
}
