% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_vector.R
\name{flip_vector}
\alias{flip_vector}
\title{Flips a vector's values and its names}
\usage{
flip_vector(x)
}
\arguments{
\item{x}{Vector to be flipped}
}
\description{
Flips a vector's values and its names
}
\details{
This function is simply a wrapper for \code{setNames(names(x), x)}
}
\examples{
x <- c(1, 2, 3)
names(x) <- c("a", "b", "c")
flip_vector(x)
}
