% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_file.R
\name{function_file}
\alias{function_file}
\title{Get the name of a file in a package where a function is defined}
\usage{
function_file(f)
}
\arguments{
\item{f}{An unquoted string with the name of the function to look up.}
}
\value{
A path to the file where the function is defined.
}
\description{
Get the name of a file in a package where a function is defined
}
\examples{
function_file(function_file)
}
