% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_lapply.R
\name{multi_lapply}
\alias{multi_lapply}
\title{Run multiple functions on a list}
\usage{
multi_lapply(...)
}
\arguments{
\item{...}{Arguments containing the list to be transformed, and functions to
apply. The first argument should be the list, and following arguments should
be the functions.}
}
\value{
A list
}
\description{
Run multiple functions on a list
}
\details{
This function is used as an alternative to piping consecutive lapplys
}
\examples{
split_cars <- split(mtcars, rep(1:3, length.out=nrow(mtcars),
each=ceiling(nrow(mtcars)/3)))
res <- multi_lapply(split_cars, t, as.data.frame)
}
