% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_rows.R
\name{name_rows}
\alias{name_rows}
\title{Re-set the row names of a data.frame with a column and delete it from the df.}
\usage{
name_rows(df, column)
}
\arguments{
\item{df}{A data.frame}

\item{column}{A character string denoting the column to be used for renaming.}
}
\value{
A data.frame with the given column as row names and without the column
}
\description{
Re-set the row names of a data.frame with a column and delete it from the df.
}
\examples{
x <- head(mtcars)
x$names <- rownames(x)
rownames(x) <- NULL
x <- name_rows(x, "names")
x
}
