% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{See the neighbors of a row in a data.frame}
\usage{
neighbors(x, row, surround = 3, colnum = 6)
}
\arguments{
\item{x}{A \code{data.frame} to subset}

\item{row}{The index of the row to see}

\item{surround}{The number of rows around 'row' to see on each side. Default: 3}

\item{colnum}{The number of columns to select. Default: 6}
}
\value{
A \code{data.frame}
}
\description{
See the neighbors of a row in a data.frame
}
\details{
This function is intended to provide a glimpse into the vicinity of a \code{data.frame}'s
row, returning the row in question, along with its \code{surround} previous and posterior
neighboring rows.
}
\examples{
neighbors(iris, 50, surround = 2)
}
