% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if.R
\name{stop_if}
\alias{stop_if}
\title{Stop execution on TRUE}
\usage{
stop_if(expr, error)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{error}{Error message to return}
}
\value{
An object of class 'error'
}
\description{
Stop execution on TRUE
}
\details{
This is a convenience function that stops the running code when
\code{expr} is \code{TRUE} and displays the text in \code{error}.
}
\examples{
# Using dontrun as the function returns an error
\dontrun{
a <- 2
stop_if(a==2, "a is equal to 2")

# This is just a short form of
if(a == 2){
    stop("a is equal to 2")
}
}
}
